multimenu

+++ server/src/api/event.ts
@@ -7,7 +7,7 @@ class EventAPI {
 
     async getAll(): Promise<Event[]> {
         return await db.query(`
-            SELECT id, name, date, status,
+            SELECT events.id, events.name name, events.date, events.status, menu.name manu_name,
             (SELECT count(tables.id) FROM tables WHERE event_id = events.id) tableCount,
             (SELECT count(items.id) FROM items WHERE event_id = events.id AND items.type = 'Bevanda') beverageCount,
             (SELECT count(items.id) FROM items WHERE event_id = events.id AND items.type IN ('Cibo')) foodCount,
@@ -16,6 +16,7 @@ class EventAPI {
             (SELECT sum(items.price) FROM items WHERE items.event_id = events.id AND type != 'Sconto' AND items.paid = 1) currentPaid,
             (SELECT count(tables.id) FROM tables WHERE tables.event_id = events.id AND tables.status = 'ACTIVE') tablesOpen
             FROM events
+            INNER JOIN menu ON events.menu_id = menu.id
             ORDER BY date DESC`
             , [])
     }
@@ -67,10 +68,14 @@ class EventAPI {
     }
 
     async create(event: Event): Promise<number> {
-        return db.executeInsert('INSERT INTO events (name, date, status) VALUES (?,?,?)', [event.name, (event.date + "").split('T')[0], 'PLANNED'])
+        return db.executeInsert('INSERT INTO events (name, date, status, menu_id) VALUES (?,?,?,?)', [event.name, (event.date + "").split('T')[0], 'PLANNED', event.menu_id])
     }
 
     async delete(id: number): Promise<number> {
+        const tables = await db.query('SELECT id FROM tables WHERE event_id = ?', [id])
+        if (tables.length) {
+            throw new Error("Can't delete the event. Tables connected")
+        }
         return await db.executeUpdate('DELETE FROM events WHERE id = ?', [id])
     }
